requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
httpx==0.27.2
python-dotenv==1.0.1
langgraph==0.2.45
langchain==0.3.11
langchain-groq==0.2.1
langchain-core==0.3.21
groq==0.11.0
jupyter==1.1.1
ipykernel==6.29.5
pandas==2.2.3
aiohttp==3.11.10
python-multipart==0.0.17
